backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
openai==2.29.0
anthropic==0.86.0
google.genai==1.68.0
python-dotenv==1.2.2
pydantic==2.12.5
python-multipart==0.0.22
pypdf==6.9.2